split into utils/extension, try to fix hover
