+++ src/extension.ts
@@ -1,9 +1,9 @@
 import path from 'path';
 import * as vscode from 'vscode';
 import { Uri } from 'vscode';
-import { format as timeagoFormat } from 'timeago.js';
-import { Blame, Change, getBlames, getChanges, getEmptyTree, getFileStatus, getGitRepository, getParentCommitId, getRepoWebBase, buildCommitUrl } from './git';
-
+import { Blame, getBlames, getChanges, getEmptyTree, getFileStatus, getGitRepository, getParentCommitId, getRepoWebBase, buildCommitUrl } from './git';
+import { buildUncommitBlame, formatDate, getCommitColor, getTextWidth, resolveChange, toMultiFileDiffEditorUris, trancateText } from './utils'
+import type { DateFormatStyle } from './utils';
 
 // 全局状态
 const fileBlameStates = new Map<string, boolean>();
@@ -16,7 +16,7 @@ const fileDecorations = new Map<string, {
     highlightDecorationType: vscode.TextEditorDecorationType | undefined,
 }>();
 
-type DateFormatStyle = 'YYYY-MM-DD' | 'Y/M/D' | 'DD.MM.YYYY' | 'relative'
+
 interface BlameDisplayConfig {
     mergeCommitLines: boolean;
     dateFormatStyle: DateFormatStyle
@@ -26,8 +26,8 @@ const MaxTitleWidth = 25;
 const VALID_FORMATS: DateFormatStyle[] = ['YYYY-MM-DD', 'Y/M/D', 'DD.MM.YYYY', 'relative'];
 
 /**
- * 激活插件
- */
+* 激活插件
+*/
 export function activate(context: vscode.ExtensionContext) {
     registerCommands(context);
     registerListeners(context);
@@ -49,8 +49,8 @@ export function activate(context: vscode.ExtensionContext) {
 }
 
 /**
- * 卸载插件
- */
+* 卸载插件
+*/
 export function deactivate() {
     for (const [_, decorations] of fileDecorations) {
         decorations.decorationTypes?.forEach(type => type.dispose());
@@ -63,8 +63,8 @@ export function deactivate() {
 
 
 /**
- * 注册命令
- */
+* 注册命令
+*/
 function registerCommands(context: vscode.ExtensionContext) {
 
     // Toggle blame annotations
@@ -134,8 +134,8 @@ function registerCommands(context: vscode.ExtensionContext) {
 }
 
 /**
- * 注册事件
- */
+* 注册事件
+*/
 function registerListeners(context: vscode.ExtensionContext) {
 
     // Editor Change
@@ -227,8 +227,8 @@ function registerListeners(context: vscode.ExtensionContext) {
 }
 
 /**
- * 显示装饰器
- */
+* 显示装饰器
+*/
 async function showDecorations(editors: vscode.TextEditor[], reload: boolean = false): Promise<boolean> {
     const document = editors[0].document;
     const documentUri = document.uri.toString();
@@ -260,8 +260,9 @@ async function showDecorations(editors: vscode.TextEditor[], reload: boolean = f
             lineBlames: undefined,
             highlightDecorationType: undefined,
         };
+
+        fileDecorations.set(documentUri, decorations);
     }
-    fileDecorations.set(documentUri, decorations);
 
     try {
         // Blames
@@ -278,7 +279,8 @@ async function showDecorations(editors: vscode.TextEditor[], reload: boolean = f
         if (!decorations.decorationTypes) {
             decorations.decorationTypes = [];
         }
-        decorations.decorationOptions = buildDecorationOptions(blames);
+
+        decorations.decorationOptions = buildDecorationOptions(blames, document.fileName, repoWebBase);
         for (const editor of editors) {
             const decorationNum = decorations.decorationOptions.length;
             for (let i = 0; i < decorationNum; i++) {
@@ -291,40 +293,7 @@ async function showDecorations(editors: vscode.TextEditor[], reload: boolean = f
         decorations.blames = blames;
         decorations.lineBlames = new Map(blames.map((blame, index) => [index, blame]));
         decorations.hoverProvider?.dispose();
-        decorations.hoverProvider = vscode.languages.registerHoverProvider(
-            { scheme: 'file', pattern: document.fileName },
-            {
-                async provideHover(document: vscode.TextDocument, position: vscode.Position) {
-                    if (position.character > 0) {
-                        return undefined;
-                    }
-                    const blame = fileDecorations.get(documentUri)?.lineBlames?.get(position.line);
-                    if (blame && blame.commited) {
-                        const rawStyle = vscode.workspace.getConfiguration('gitblame').get<string>('dateFormatStyle', 'relative');
-                        const activeStyle: DateFormatStyle = VALID_FORMATS.includes(rawStyle as DateFormatStyle)
-                            ? (rawStyle as DateFormatStyle)
-                            : 'relative';
-                        const hoverStyle: DateFormatStyle = activeStyle === 'relative' ? 'YYYY-MM-DD' : activeStyle;
-                        const dateText = formatDate(blame.timestamp, hoverStyle);
-
-                        const content = new vscode.MarkdownString();
-                        const [commitUrl, gitPlatform] = buildCommitUrl(repoWebBase, blame.commit);
-                        if (commitUrl) {
-                            const viewText = gitPlatform ? `View on ${gitPlatform}` : 'Open in Browser';
-                            content.appendMarkdown(`[${viewText}](${commitUrl})  \n`);
-                        }
-                        content.appendMarkdown(`commit: [${blame.commit}](command:git.blame.viewCommit?${encodeURIComponent(JSON.stringify([blame.commit, blame.summary, document.fileName]))})  \n`);
-                        content.appendMarkdown(`Author: ${blame.author}  \n`);
-                        content.appendMarkdown(`Date: ${dateText}  \n`);
-                        if (blame.summary) {
-                            content.appendMarkdown(`\n\n${blame.summary}`);
-                        }
-                        content.isTrusted = true;
-                        return new vscode.Hover(content);
-                    }
-                }
-            }
-        );
+        decorations.hoverProvider = undefined;
         fileBlameStates.set(documentUri, true);
         return true;
     } catch (error: any) {
@@ -338,8 +307,8 @@ async function showDecorations(editors: vscode.TextEditor[], reload: boolean = f
 
 
 /**
- * 隐藏装饰器
- */
+* 隐藏装饰器
+*/
 async function hideDecorations(document: vscode.TextDocument): Promise<boolean> {
     const documentUri = document.uri.toString();
     fileBlameStates.set(documentUri, false);
@@ -355,8 +324,8 @@ async function hideDecorations(document: vscode.TextDocument): Promise<boolean> 
 }
 
 /**
- * 更新装饰器
- */
+* 更新装饰器
+*/
 async function updateDecorationsOnChange(editors: vscode.TextEditor[], event: vscode.TextDocumentChangeEvent) {
     const documentUri = editors[0].document.uri.toString();
     const decorations = fileDecorations.get(documentUri);
@@ -375,6 +344,7 @@ async function updateDecorationsOnChange(editors: vscode.TextEditor[], event: vs
         if (addedLines.length === 0 && deletedLines.length === 0 && modifiedLines.length === 0) {
             continue;
         }
+
         if (modifiedLines.length > 0) {
             for (let i = 0; i < modifiedLines.length; i++) {
                 if (blames[modifiedLines[i]].commited) {
@@ -384,12 +354,14 @@ async function updateDecorationsOnChange(editors: vscode.TextEditor[], event: vs
                 }
             }
         }
+
         if (deletedLines.length > 0) {
             shouldUpdate = true;
             for (let i = deletedLines.length - 1; i >= 0; i--) {
                 blames.splice(deletedLines[i], 1);
             }
         }
+
         if (addedLines.length > 0) {
             shouldUpdate = true;
             for (let i = 0; i < addedLines.length; i++) {
@@ -397,29 +369,34 @@ async function updateDecorationsOnChange(editors: vscode.TextEditor[], event: vs
             }
         }
     }
+
     if (!shouldUpdate) {
         return;
     }
 
+    const repositoryRoot = await getGitRepository(editors[0].document.fileName);
+    const repoWebBase = repositoryRoot ? await getRepoWebBase(repositoryRoot) : "";
+
     // update decorations
-    decorations.decorationOptions = buildDecorationOptions(blames);
+    decorations.decorationOptions = buildDecorationOptions(blames, editors[0].document.fileName, repoWebBase);
     for (const editor of editors) {
         const decorationNum = decorations.decorationOptions.length;
         for (let i = 0; i < decorationNum; i++) {
             if (i >= decorations.decorationTypes.length) {
                 decorations.decorationTypes.push(vscode.window.createTextEditorDecorationType({}));
             }
+
             editor.setDecorations(decorations.decorationTypes[i], decorations.decorationOptions[i]);
         }
     }
+
     decorations.blames = blames;
     decorations.lineBlames = new Map(blames.map((blame, index) => [index, blame]));
 }
 
-
 /**
- * 更新上下文菜单
- */
+* 更新上下文菜单
+*/
 async function updateMenuContext(document: vscode.TextDocument, currentState: boolean | undefined = undefined) {
     // Skip diff editor
     if (document.uri.scheme !== 'file') {
@@ -443,6 +420,7 @@ async function updateMenuContext(document: vscode.TextDocument, currentState: bo
             vscode.commands.executeCommand('setContext', 'gitblame.hideMenuState', false);
             return;
         }
+
         // check file blame state
         const fileBlameState = fileBlameStates.get(document.uri.toString());
         vscode.commands.executeCommand('setContext', 'gitblame.showMenuState', !fileBlameState);
@@ -452,19 +430,18 @@ async function updateMenuContext(document: vscode.TextDocument, currentState: bo
         vscode.commands.executeCommand('setContext', 'gitblame.showMenuState', false);
         vscode.commands.executeCommand('setContext', 'gitblame.hideMenuState', false);
     }
-
 }
 
-function buildDecorationOptions(blames: Blame[]): vscode.DecorationOptions[][] {
+function buildDecorationOptions(blames: Blame[], fileName: string, repoWebBase: string): vscode.DecorationOptions[][] {
     const cfg = vscode.workspace.getConfiguration('gitblame');
 
-    const rawDateFormatStyle = cfg.get<string>('dateFormatStyle', 'relative');
+    const rawDateFormatStyle = cfg.get('dateFormatStyle', 'relative');
     const validDateFormatStyle = VALID_FORMATS.includes(rawDateFormatStyle as DateFormatStyle)
         ? (rawDateFormatStyle as DateFormatStyle)
         : 'relative';
 
     const config: BlameDisplayConfig = {
-        mergeCommitLines: cfg.get<boolean>('mergeCommitLines', true),
+        mergeCommitLines: cfg.get('mergeCommitLines', true),
         dateFormatStyle: validDateFormatStyle
     };
 
@@ -472,6 +449,7 @@ function buildDecorationOptions(blames: Blame[]): vscode.DecorationOptions[][] {
     if (maxWidth <= 0) {
         return [];
     }
+
     const singleCommit = new Set(blames.filter(b => b.commited).map(b => b.commit)).size === 1;
 
     const decorationOptions: vscode.DecorationOptions[] = [];
@@ -483,12 +461,15 @@ function buildDecorationOptions(blames: Blame[]): vscode.DecorationOptions[][] {
             color = getCommitColor(blame.commit, blame.timestamp);
             colorsMap.set(blame.commit, color);
         }
+
         const range = new vscode.Range(
             new vscode.Position(index, 0),
             new vscode.Position(index, 0)
         );
-        const option: { range: vscode.Range, renderOptions: vscode.DecorationRenderOptions } = {
+        const hoverMessage = buildHoverMessage(blame, fileName, repoWebBase);
+        const option: vscode.DecorationOptions = {
             range,
+            hoverMessage,
             renderOptions: {
                 before: {
                     contentText: `\u2007${blame.title}\u2007`,
@@ -501,8 +482,9 @@ function buildDecorationOptions(blames: Blame[]): vscode.DecorationOptions[][] {
         };
         decorationOptions.push(option);
 
-        const optionHeatmap = {
+        const optionHeatmap: vscode.DecorationOptions = {
             range,
+            hoverMessage,
             renderOptions: {
                 before: {
                     contentText: '\u2007',
@@ -522,8 +504,8 @@ function buildDecorationOptions(blames: Blame[]): vscode.DecorationOptions[][] {
             }
         };
         if (singleCommit || !blame.commited) {
-            optionHeatmap.renderOptions.light.before.backgroundColor = 'transparent';
-            optionHeatmap.renderOptions.dark.before.backgroundColor = 'transparent';
+            optionHeatmap.renderOptions!.light!.before!.backgroundColor = 'transparent';
+            optionHeatmap.renderOptions!.dark!.before!.backgroundColor = 'transparent';
         }
         decorationOptionsHeatmap.push(optionHeatmap);
     });
@@ -531,6 +513,35 @@ function buildDecorationOptions(blames: Blame[]): vscode.DecorationOptions[][] {
     return [decorationOptions, decorationOptionsHeatmap];
 }
 
+function buildHoverMessage(blame: Blame, fileName: string, repoWebBase: string): vscode.MarkdownString | undefined {
+    if (!blame.commited) {
+        return undefined;
+    }
+
+    const rawStyle = vscode.workspace.getConfiguration('gitblame').get('dateFormatStyle', 'relative');
+    const activeStyle: DateFormatStyle = VALID_FORMATS.includes(rawStyle as DateFormatStyle)
+        ? (rawStyle as DateFormatStyle)
+        : 'relative';
+    const hoverStyle: DateFormatStyle = activeStyle === 'relative' ? 'YYYY-MM-DD' : activeStyle;
+    const dateText = formatDate(blame.timestamp, hoverStyle);
+
+    const content = new vscode.MarkdownString();
+    const [commitUrl, gitPlatform] = buildCommitUrl(repoWebBase, blame.commit);
+    if (commitUrl) {
+        const viewText = gitPlatform ? `View on ${gitPlatform}` : 'Open in Browser';
+        content.appendMarkdown(`[${viewText}](${commitUrl}) \n`);
+    }
+
+    content.appendMarkdown(`commit: [${blame.commit}](command:git.blame.viewCommit?${encodeURIComponent(JSON.stringify([blame.commit, blame.summary, fileName]))}) \n`);
+    content.appendMarkdown(`Author: ${blame.author} \n`);
+    content.appendMarkdown(`Date: ${dateText} \n`);
+    if (blame.summary) {
+        content.appendMarkdown(`\n\n${blame.summary}`);
+    }
+
+    content.isTrusted = true;
+    return content;
+}
 
 function fillTitles(blames: Blame[], config: BlameDisplayConfig): number {
     let maxWidth = 0;
@@ -580,224 +591,4 @@ function fillTitles(blames: Blame[], config: BlameDisplayConfig): number {
     }
 
     return maxWidth;
-}
-
-// ------------------------------------------------------------
-// utils
-// ------------------------------------------------------------
-/** 
- * @param timestamp timestamp in seconds
- */
-function formatDate(timestamp: number, style: DateFormatStyle): string {
-    if (style === 'relative') {
-        const locale = vscode.env.language.startsWith('zh') ? 'zh_CN' : 'en_US';
-        return timeagoFormat(new Date(timestamp * 1000), locale);
-    }
-
-    const date = new Date(timestamp * 1000);
-    const map: Record<string, [string, Intl.DateTimeFormatOptions]> = {
-        'YYYY-MM-DD': ['en-CA', { year: 'numeric', month: '2-digit', day: '2-digit' }],
-        'Y/M/D': ['en-GB', { year: 'numeric', month: 'numeric', day: 'numeric' }],
-        'DD.MM.YYYY': ['de-DE', { year: 'numeric', month: '2-digit', day: '2-digit' }],
-    };
-    const [locale, options] = map[style];
-    return new Intl.DateTimeFormat(locale, options).format(date);
-}
-
-function toMultiFileDiffEditorUris(change: Change, originalRef: string, modifiedRef: string): { originalUri: Uri | undefined; modifiedUri: Uri | undefined } {
-    switch (change.status) {
-        case "index_added":
-            return {
-                originalUri: undefined,
-                modifiedUri: toGitUri(change.uri, modifiedRef)
-            };
-        case "deleted":
-            return {
-                originalUri: toGitUri(change.uri, originalRef),
-                modifiedUri: undefined
-            };
-        case "index_renamed":
-            return {
-                originalUri: toGitUri(change.originalUri, originalRef),
-                modifiedUri: toGitUri(change.uri, modifiedRef)
-            };
-        default:
-            return {
-                originalUri: toGitUri(change.uri, originalRef),
-                modifiedUri: toGitUri(change.uri, modifiedRef)
-            };
-    }
-}
-
-function toGitUri(uri: Uri, ref: string, options: { submoduleOf?: string, replaceFileExtension?: boolean, scheme?: string } = {}): Uri {
-    const params = {
-        path: uri.fsPath,
-        submoduleOf: "",
-        ref
-    };
-
-    if (options.submoduleOf) {
-        params.submoduleOf = options.submoduleOf;
-    }
-
-    let path = uri.path;
-
-    if (options.replaceFileExtension) {
-        path = `${path}.git`;
-    } else if (options.submoduleOf) {
-        path = `${path}.diff`;
-    }
-
-    return uri.with({ scheme: options.scheme ?? 'git', path, query: JSON.stringify(params) });
-}
-
-
-function getTextWidth(text: string): { width: number, widths: number[] } {
-    let width = 0;
-    const widths = [];
-    for (const char of text) {
-        const w = getCharacterWidth(char);
-        widths.push(w);
-        width += w;
-    }
-    return { width, widths };
-}
-
-
-function getCharacterWidth(char: string): number {
-    const code = char.charCodeAt(0);
-
-    // 东亚文字 (中文、日文、韩文等)
-    if ((code >= 0x3000 && code <= 0x9FFF) ||
-        (code >= 0xAC00 && code <= 0xD7AF) ||
-        (code >= 0xF900 && code <= 0xFAFF) ||
-        (code >= 0xFF00 && code <= 0xFFEF)) {
-        return 2;
-    }
-
-    // 表情符号和特殊符号
-    if (code >= 0x1F300 && code <= 0x1F9FF) {
-        return 2;
-    }
-
-    // 组合字符标记
-    if (code >= 0x0300 && code <= 0x036F) {
-        return 0;
-    }
-
-    return 1;
-}
-
-function trancateText(text: string, maxWidth: number, widths: number[]): string {
-    let truncatedText = '';
-    let currentWidth = 0;
-
-    for (let i = 0; i < widths.length; i++) {
-        if (currentWidth + widths[i] <= maxWidth) {
-            truncatedText += text[i];
-            currentWidth += widths[i];
-        } else {
-            break;
-        }
-    }
-    return truncatedText;
-}
-
-function getCommitColor(commit: string, timestamp: number): { lightColor: string, darkColor: string } {
-    // hue
-    let hash = 0;
-    for (let i = 0; i < commit.length; i++) {
-        hash = commit.charCodeAt(i) + ((hash << 5) - hash);
-    }
-    const h = (hash * 137.508) % 360;
-
-    // saturation
-    const minSaturation = 35, maxSaturation = 90, decayDays = 20;
-    let daysAgo = Math.floor((Date.now() / 1000 - timestamp) / (24 * 60 * 60));
-    daysAgo = Math.max(daysAgo, 0);
-    const decay = Math.exp(-daysAgo / 20);
-    const saturation = Math.round(minSaturation + (maxSaturation - minSaturation) * decay);
-
-    const darkColor = `hsl(${h}, ${saturation}%, 50%)`;
-    const lightColor = `hsl(${h}, ${saturation}%, 50%)`;
-    return { lightColor, darkColor };
-}
-
-function buildUncommitBlame(line: number): Blame {
-    return {
-        line: line,
-        commit: '0000000000000000000000000000000000000000',
-        author: '',
-        mail: '',
-        timestamp: 0,
-        summary: '',
-        commited: false,
-        title: '',
-    };
-}
-
-
-function resolveChange(change: vscode.TextDocumentContentChangeEvent) {
-    const addedLines = [];
-    const deletedLines = [];
-    const modifiedLines = [];
-    const changeText = change.text;
-    const startLine = change.range.start.line;
-    const endLine = change.range.end.line;
-    const startLineCharacter = change.range.start.character;
-
-    if (changeText.length === 0) {
-        // delete characters
-        const diffLine = endLine - startLine;
-        if (diffLine === 1) {
-            deletedLines.push(startLine + 1);
-        } else if (diffLine > 1) {
-            const start = startLineCharacter > 0 ? startLine + 1 : startLine;
-            const end = start + diffLine - 1;
-            for (let i = start; i <= end; i++) {
-                deletedLines.push(i);
-            }
-        } else if (diffLine === 0) {
-            modifiedLines.push(startLine);
-        }
-    } else {
-        const trimedChangeText = changeText.replace(/ +$/, '');
-        if (trimedChangeText === '\n' || trimedChangeText === '\r\n') {
-            // add a new line
-            addedLines.push(startLineCharacter > 0 ? startLine + 1 : startLine);
-        } else {
-            // add or modify characters
-            const crossLines = endLine - startLine + 1;
-            const textLines = changeText.split(/\r?\n/).length;
-            const diff = textLines - crossLines;
-            if (diff > 0) {
-                // modify lines
-                for (let i = startLine; i <= endLine; i++) {
-                    modifiedLines.push(i);
-                }
-                // add lines
-                const start = endLine + 1;
-                const end = endLine + diff;
-                for (let i = start; i <= end; i++) {
-                    addedLines.push(i);
-                }
-            } else if (diff < 0) {
-                // modify lines
-                for (let i = startLine; i <= endLine + diff; i++) {
-                    modifiedLines.push(i);
-                }
-                const start = endLine + diff + 1;
-                const end = endLine;
-                for (let i = start; i <= end; i++) {
-                    deletedLines.push(i);
-                }
-            } else if (diff === 0) {
-                // modify lines
-                for (let i = startLine; i <= endLine; i++) {
-                    modifiedLines.push(i);
-                }
-            }
-        }
-    }
-    return { addedLines, deletedLines, modifiedLines };
 }
